HCOLL-172 Verified thath IntIntMultiMap now supports up to 2^32 entries; ShortShortMultiMap up to 2^16; ChronicleMap.Segment are now long-sized; attempt to support non-power-of-2 number of segments by HashSplitting (not tested yet); improved logic and dependencies between actualSegments(), actualEntriesPerSegments() and minSegments() methods in ChronicleMapBuilder; minor fixes

+++ src/main/java/net/openhft/chronicle/map/MapEventListener.java
@@ -95,7 +95,7 @@ public abstract class MapEventListener<K, V, M extends ChronicleMap<K, V>>
     }
 
     void onRemove(M map, Bytes entry, int metaDataBytes, K key, V value,
-                  int pos, SharedSegment segment) {
+                  long pos, SharedSegment segment) {
         onRemove(map, entry, metaDataBytes, key, value);
     }
 
@@ -112,7 +112,7 @@ public abstract class MapEventListener<K, V, M extends ChronicleMap<K, V>>
         // do nothing
     }
 
-    void onRelocation(int pos, SharedSegment segment) {
+    void onRelocation(long pos, SharedSegment segment) {
         // do nothing
     }
 }
